LayoutTests/js/eval-var-decl-expected.txt

This test case checks whether variables cause properties to be defined even before reaching the declaration statement in various cases.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS this.hasOwnProperty("foo") is true
PASS this.hasOwnProperty("bar") is true
PASS firstEvalResult is true
PASS secondEvalResult is false
PASS thirdEvalResult is true
PASS testEvalInCatch() is true
PASS successfullyParsed is true

TEST COMPLETE
